resolve deprecation and other warnings per Xcode 8
rm unneeded (now incorrect) orientation checks; corrected todo-txt web address in Settings page

Xcode 9 warning/deprecation resolution

fix more deprecation warnings

Update Pickers library

From https://github.com/skywinder/ActionSheetPicker-3.0, the fork
recommended by the original author

+++ Classes/TaskBagImpl.m
@@ -222,7 +222,7 @@ static Task* find(NSArray *tasks, Task *task) {
 	return localTasks;
 }
 
-- (int) size {
+- (NSUInteger) size {
     return [self.tasks count];
 }
